+++ apps/account-server/src/lib/formatting.ts
@@ -1,10 +1,3 @@
-export function shortenAddress(
-  address: `0x${string}` | string | undefined,
-  chars = 3,
-): string | undefined {
-  return address && `${address.slice(0, chars + 3)}...${address.slice(-5)}`;
-}
-
 export function formatNumberWithCommas(number: number): string {
   // Split number into integer and decimal parts
   const parts = number.toString().split('.');